+++ demo/test_import.py
@@ -1,8 +1,0 @@
-import nuke
-import hiero
-
-
-
-nuke.createNode()
-
-hiero.core.newProject()