+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [0.14.0] - 02/18/2024
+
+### Changed
+
+- Added a monthly check for the latest included packages.
+- Changed the command `Nuke: Add Pyside2 Plugin` functionality to reflect the new changes in the included template.
+
 ## [0.13.0] - 03/12/2023
 
 ### Added
@@ -10,7 +17,7 @@
 
 ### Changed
 
-- Update NukeServerSocket to version to latest
+- Update nukeserversocket to version to latest
 - Update socket debug functionallity.
 
 ## [0.12.0] - 09/23/2023
@@ -118,7 +125,7 @@
 
 ### Added
 
-- New command that adds NukeServerSocket plugin inside Nuke's directory (`.nuke`) & `menu.py`.
+- New command that adds nukeserversocket plugin inside Nuke's directory (`.nuke`) & `menu.py`.
 
 ## [0.4.5] 02/18/2022
 
@@ -206,7 +213,7 @@ Mostly code refactoring and test suite.
 
 ### Fixed
 
-- Fixed file name reference that didn't allow vscode to pick NukeServerSocket port settings automatically.
+- Fixed file name reference that didn't allow vscode to pick nukeserversocket port settings automatically.
 
 ## [0.0.2] - [0.0.8]